feat: error rate + welcome message (#171)
This PR adds information and guiding UX on the sequencing error rate.
That is we don't show mutations under the sequencing error rate anymore.
With that, we also show information on the processing that is done on
the index page.

Further, we changed the welcome graphic that is shown to make the
message less technical, i.e. this project was hard, but rather this is
what you can do with it.

Co-authored-by: Gordon J. Köhn <gordon.koehn@dbsse.ethz.ch>